+++ spec/rails_app/db/migrate/20150217000000_create_tables.rb
@@ -1,4 +1,10 @@
-class CreateTables < ActiveRecord::Migration
+if Rails::VERSION::MAJOR > 4
+  migration_class = ActiveRecord::Migration[4.2]
+else
+  migration_class = ActiveRecord::Migration
+end
+
+class CreateTables < migration_class
   def change
     create_table(:users) do |t|
       ## Database authenticatable
@@ -15,6 +21,8 @@ class CreateTables < ActiveRecord::Migration
       t.datetime :last_sign_in_at
       t.string   :current_sign_in_ip
       t.string   :last_sign_in_ip
+
+      t.string :nick_name
 
       t.timestamps null: false
     end